CRIMSON: SoC code running on Crimson.
This is the code for running the server on Per Vices Crimson devices.
The server handles UDP requests and file i/o within the main folder of
the SoC ~/state/*. The server will handle those requests and in return
trigger the correct MCU command, or FPGA fabric memory transaction.

The code is cross compiled with arm-linux-gnueabihf-gcc. The files will
be built in an output directory with all the source, header, object, and
executable files.