Modify dither_sra_sel input range.
Previously, the handlers for dither_sra_sel for each channel would accept input in the range [0..15].

That range was not monotonically increasing, since 0 represents the maximum and 1 represents the minimum.

A better input range would be [1..16], which is monotonically increasing. The correct register values can be
obtained by a % 16 operation.

However, each step is actually 6 dB, so the (human readable) range is [6..96].

The handlers were therefore modified to accept values in the range [6..96] in multiples of 6. The value passed
in is then divided by 6, and goes through a % 16 operation to get the appropriate register value.