PLL_CALC: Moving to fixed PFD integer N design.
This commit refactors a lot of code to move the design
to a fixed, integer-N, architecture. While this limits
the tunable frequencies that we can hit to multiples of
the PFD frequency, it greatly simplifies a lot of the
calculations.

It also makes it easier for us to migrate to a known
phase coherency once we're able to set up the LMKs for
that.