Refactored logging
Dump file was preserved. All other PRINT priorities write to stdout and stderr.
This was done to simplify logging. Systemd will buffer the logs and truncate the logs if they become too large.
systemd also preserves the log of a given process in case you need to get an error report.