Removing extsine and commenting out temp property
Removing extsine: This was in Vaunt but not for Cyan, causes unregonized
command error.

Commenting out Temp property: Due to it not being implemented correctly,
this comand causes the system to reboot

Changes on server.c comments out anything to do with TX since currently,
it is not needed and is causing a longer boot time.