Changing the way that the system decides to enable boards
The old method checked if the path contained rx or tx, then set pwr for
the repective board to 1. That method causes difficulties when trying to
run a function that is meant to control part of power. The fix is to add
a flag to the property that says what whether or not to turn on a board.
Anything not specified by the flag will run using the old method