TATE SERVER 4r4t_3g final jesd_error property
Created and tested a property to reset the jesd error codes. This will
try and reset all rx possibilities (up to 16 Rx boards in a unit).
Therefore for each server change we do not need to adjust the jesd_rx[]
array to the number of RX featured.
Currently only on the 4r4t_3g server. Will be adding to the other
servers soon.