PROPERTY_MANAGER: removing setting buffer to 0 when checking property
Previously the code would set the entire read buffer to 0 when checking
any property. Theoretically this is useful because it prevents stale
data from affecting anything else. However, it is very time consuming
with large buffers and has been removed. This should no have any impact
unless there is a bug elsewhere in the code