VAUNT: improving JESD bringup sequence
-FPGA reset would bring down tx JESD. Presumably something done during
sync_channels_prep would be reset during the FPGA reset. The fix is to
unseperate sync_channels and run the whole thing after the FPGA reset

-Removed the JESD sync during channel power on since it is no longer
needed. Also avoids the risk of a race condition on the first run
resyncing JESD on channels in already in use when powering on a
different channel