TATE,USE_3G_AS_1G: get RX HW ver from fw_ver prop
The USE_3G_AS_1G flag is used to specify that the backplane is 3G, but the RF chain is 1G. We are now supporting the RF chain as 3G configured as 1G or simply as 1G. There is one difference between these cases: the 3G configured as 1G has a 250MHz nco already. This affects how we set the FPGA NCO, so we will detect which hardware version the RX board is from the /about/fe_ver property.