TATE/LILY: change state tree to volatile
This was planned for a long time to reduce wear on the SD card and
improve performance. Theoretically this should speed up reads to the
state tree.

I am doing this now to avoid issue where the state must be manually
deleted if moving between severs that change a path between files and
links