CYAN: implementing feedback registers
Added feedback registers for SFP and JESD debug info.
The bits are intended to be high when errors are detected. The first
implementation of these in the FPGA has the bits be high to indicate
good so for a few days until the the swap is made in the FPGA it will
show the opposite of the actual result