AVERY: adding hardware/product version decoding
First step of improving the way Avery is organized. Currently it is set
up as a variant of Crimson, going forward it is going to be set up the
same way Chestnut is, but based off of Crimson instead of Cyan.

Also corrected properties_vaunt.c so that it is used HARDWARE_RTM_VER instead of
PRODUCT_RTM_VER